feat: 기존 주석 처리된 rent 연동 기능 활성화

+++ backend/bookbook/src/main/java/com/bookbook/domain/wishList/dto/WishListCreateRequestDto.java
@@ -1,6 +1,6 @@
 package com.bookbook.domain.wishList.dto;
 
 public record WishListCreateRequestDto(
-        Long rentId
+        Integer rentId
 ) {
 }
